Add initial implementation and tests.
This initial version gets the core functionality for a source connector working
for a single table, periodically polling, reading rows, translating to Avro, and
passing them to the framework for storage. It's missing any sort of incremental
load, doesn't have complete coverage of data types, doesn't pick up changes to
the set of tables automatically, etc.