Add SQLite autoincrement example config and handle missing metadata in SQLite driver.
The SQLite JDBC driver doesn't include all the metadata that it is supposed to
return. One of the things it omits is the autoincrement property, so a fallback
that selects one row of data and uses the ResultSetMetaData to determine this
property is required. Also add some trace-level logging to the JDBCSourceTask to
aid debugging.