Add a check for the current time when querying using timestamps.
This is necessary because the timestamp modes need to make sure all data for a
timestamp has been written before returning any data with that
timestamp. Without this check, data might be returned, new data with the same
timestamp written, and a subsequent query would not pick it up. In some
databases, the granularity will be quite large, e.g. 1s, so this is actually
very likely to be a problem.