Add check for NOT NULL columns on target tables when specified as increasing or timestamp column for offsets.
Because this is checked before making any queries, this is only checked for
table-based copying. Custom queries would either require specifying a
table.column format for the check, which wouldn't guarantee it even matched up
with the query, or require parsing or executing the query to perform the
validation before normal polling could start.