CC-1065 Source connector’s recommender uses table type filter and caches results
The JDBC source connector’s recommender for table names should not perform better for databases that have large numbers of tables. The recommender now uses the table types to reduce the number of results that must be processed when determining the names of the available tables. More importantly, each ConfigDef instance returned by the connector’s method will use a table name recommender that caches the results for a given configuration, so using the same recommender instance only queries the tables once and returns the same list of table names for both the whitelist and blacklist configuration.