CC-1193: Refactored with a new dialect framework for DBMS-specific behavior
The source and sink JDBC connectors were refactored to use dialects for all interaction with the database, allowing for the customization of behavior when working with specific DBMSes. Dialects are extensible and auto-discovered via Java’s service loader mechanism. Separate dialects were added for each major DBMS, though only DBMS-specific behavior already in the connector was extracted into those dialects; additional customization can be done on the dialects to handle DBMS-specific issues.

Prior to this change, the sink connector made use of an earlier dialect framework that was not very extensible, and it only encapsulated a limited subset of the interaction with the database. This new dialect framework is completely extensible (dialects can be added at runtime by simply including in the plugin or on the classpath one or more JARs with the custom dialects), and it completely encapsulates all interaction with the database so that dialects can customize any of the behavior.

A generic dialect is included as a complete implementation that uses JDBC and standard SQL, and custom dialects should extend this generic implementation so that the dialect overrides only the behavior that needs to be customized. The majority of the code within a dialect is needed by the source and sink logic, so there is no longer a distinction between source and sink dialects; each dialect supports all features of both source and sink connectors.

The old sink dialect framework was removed (rather than just deprecated), since it was difficult for users to provide their own custom sink dialect implementations, as the method that determined which sink dialect to use was static and thus not easily customized. Also refactored the sink’s database metadata queries and utilities to reuse components already used in the source.