CC-1266: Corrected build by changing PowerMock uses
A recent change to the Kafka dependencies included an upgrade of the PowerMock library to 2.0.0-beta5 for JDK9 reasons, but this appears to have broken several tests.

Apparently, it is better to use `@Mock` annotations on member fields rather than to manually create the (local or member field) mock instances via `PowerMock.createMock(MyClass.class)`. Unfortunately this requires running the concrete test classes via `@RunWith(PowerMockRunner.class)` and several other class-level annotations, which is more verbose.