CC-1193: Added preliminary Sybase dialect
Added a preliminary Sybase dialect. Note that Sybase and SQL Server are the first DBMSes supported by the connector that use both JDBC’s catalog and schema.  In fact, both use JDBC's catalog to represent the database, and JDBC's schema to represent the owner (e.g., "dbo”). However, this required moving more functionality into the dialect and required changing the source and sink connector to do more with TableId rather than just table name to properly distinguish between tables in different databases.

Also, prior to this commit the connectors assumed JDBC 4 compliant drivers. The JTDS driver for Sybase and SQL Server is JDBC 3 compliant, and does not support methods added with JDBC 4. Therefore, several changes were made to dynamically determine the driver compliance level and to call these JDBC 4 methods only when the driver is JDBC 4 compliant. And, this also allows a dialect to be less dependent upon the JDBC driver compliance level. For example, while the aforementioned JTDS driver is JDBC 3 compliant while the JConnect4 driver from Sybase is JDBC 4 compliant.

A number of changes to the dialects were necessary to handle Sybase, in part because of the very different DDL but also because of some characteristics of the JTDS driver that does not allow DDL statements to be executed within a multi-statement transaction. The latter means that the sink connector’s explicit use of disabling autocommit caused problems with this driver. Also, the PreparedStatementBinder’s logic of binding different types via JDBC methods did not allow overriding the method used to set the value, so this logic was moved into the database dialect.