CC-2139: Moved the call to set the fetch direction out of generic dialect
Setting the fetch direction does not work with some versions of SQLite and potentially other DBMSes. Since the 4.1.1 source connector did not use this method, the generic dialect should not set the fetch direction and instead only the PostgreSQL and MySQL dialects should do this since those drivers respect and use this to reduce memory problems.

The most straightforward way to handle this is to create a new protected method in `GenericDatabaseDialect` that is to perform any operations on a newly-created prepared statement, and by default to do nothing. Then the `PostgreSqlDatabaseDialect` and `MySqlDatabaseDialect` can override this and call the `setFetchDirection` method.