CC-2910: Changed source task to close connections when finished
The JDBC source task was attempting to close the CachedConnectionProvider and its connections when the task was stopped, which happens asynchronously while the `poll()` method may be still working in the task’s primary thread. If the connection provider is closed but is needed again, it will reopen connections as needed but these will never be closed.

With this change, the connection provider is closed at the end of `poll()` after the task has been stopped and after the connections are no longer being used.