Sleep in chunks
For longer polling durations, the task will spend most of its time
asleep.  If the task is asleep when it is getting killed, such as when a
rebalance is happening or the config is changed, the task is likely not
to return within the configured graceful timeout period, instead causing
`Graceful stop of task connector_name-0 failed.` errors.

This tweaks the connector to sleep in chunks of no more than 100ms in
order to keep it responsive to any changes.

Fix log line

Fix checkstyle violations

Sleep `toSleepMs`