Fix connection check in Sqlite test
The Sqlite test use the DB path not writable in normal circumstance (`/...`). This commit fixes this.

As the path becomes writable, the proper JDBC connection to Sqlite can be established. Now, the dialect will get `"` as the quotation for identifiers (preferable) instead of \` (some compatibility mode with MySQL). For this, expected values in tests need fixing.